+++ hutool-cache/src/main/java/cn/hutool/cache/impl/CacheValuesIterator.java
@@ -1,7 +1,6 @@
 package cn.hutool.cache.impl;
 
 import java.util.Iterator;
-import java.util.NoSuchElementException;
 
 /**
  *  {@link cn.hutool.cache.impl.AbstractCache} 的值迭代器.
@@ -11,17 +10,15 @@ import java.util.NoSuchElementException;
  */
 public class CacheValuesIterator<V> implements Iterator<V> {
 
-	private final Iterator<? extends CacheObj<?, V>> iterator;
-	private CacheObj<?,V> nextValue;
+	private final CacheObjIterator<?, V> cacheObjIter;
 
 	/**
 	 * 构造
-	 * @param iterator 原{@link Iterator}
+	 * @param iterator 原{@link CacheObjIterator}
 	 * @param readLock 读锁
 	 */
-	CacheValuesIterator(Iterator<? extends CacheObj<?, V>> iterator) {
-		this.iterator = iterator;
-		nextValue();
+	CacheValuesIterator(CacheObjIterator<?, V> iterator) {
+		this.cacheObjIter = iterator;
 	}
 
 	/**
@@ -29,7 +26,7 @@ public class CacheValuesIterator<V> implements Iterator<V> {
 	 */
 	@Override
 	public boolean hasNext() {
-		return nextValue != null;
+		return this.cacheObjIter.hasNext();
 	}
 
 	/**
@@ -37,33 +34,14 @@ public class CacheValuesIterator<V> implements Iterator<V> {
 	 */
 	@Override
 	public V next() {
-		if (false == hasNext()) {
-			throw new NoSuchElementException();
-		}
-		final V cachedObject = nextValue.obj;
-		nextValue();
-		return cachedObject;
+		return cacheObjIter.next().getValue();
 	}
 
 	/**
-	 * 从缓存中移除没有过期的当前值
+	 * 从缓存中移除没有过期的当前值，不支持此方法
 	 */
 	@Override
 	public void remove() {
-//		iterator.remove();
-		throw new UnsupportedOperationException("Cache values Iterator is not support to modify.");
-	}
-	
-	/**
-	 * 下一个值，当不存在则下一个值为null
-	 */
-	private void nextValue() {
-		while (iterator.hasNext()) {
-			nextValue = iterator.next();
-			if (nextValue.isExpired() == false) {
-				return;
-			}
-		}
-		nextValue = null;
+		cacheObjIter.remove();
 	}
 }