+++ hutool-core/src/main/java/cn/hutool/core/util/CharUtil.java
@@ -231,6 +231,34 @@ public class CharUtil {
 	}
 
 	/**
+	 * 是否空白符<br>
+	 * 空白符包括空格、制表符、全角空格和不间断空格<br>
+	 * 
+	 * @param c 字符
+	 * @return 是否空白符
+	 * @see Character#isWhitespace(int)
+	 * @see Character#isSpaceChar(int)
+	 * @since 4.0.10
+	 */
+	public static boolean isBlankChar(char c) {
+		return isBlankChar((int) c);
+	}
+	
+	/**
+	 * 是否空白符<br>
+	 * 空白符包括空格、制表符、全角空格和不间断空格<br>
+	 * 
+	 * @see Character#isWhitespace(int)
+	 * @see Character#isSpaceChar(int)
+	 * @param c 字符
+	 * @return 是否空白符
+	 * @since 4.0.10
+	 */
+	public static boolean isBlankChar(int c) {
+		return Character.isWhitespace(c) || Character.isSpaceChar(c) || c == '\ufeff' || c == '\u202a';
+	}
+
+	/**
 	 * 判断是否为emoji表情符<br>
 	 * 参考：http://blog.csdn.net/a445020593/article/details/56667654
 	 *