+++ hutool-core/src/main/java/cn/hutool/core/util/NumberUtil.java
@@ -1774,12 +1774,14 @@ public class NumberUtil {
 	 * 是否空白符<br>
 	 * 空白符包括空格、制表符、全角空格和不间断空格<br>
 	 * 
-	 * @see Character#isWhitespace(int)
-	 * @see Character#isSpaceChar(int)
 	 * @param c 字符
 	 * @return 是否空白符
+	 * @see Character#isWhitespace(int)
+	 * @see Character#isSpaceChar(int)
 	 * @since 3.0.6
+	 * @deprecated 请使用{@link CharUtil#isBlankChar(char)}
 	 */
+	@Deprecated
 	public static boolean isBlankChar(char c) {
 		return isBlankChar((int) c);
 	}
@@ -1793,9 +1795,11 @@ public class NumberUtil {
 	 * @param c 字符
 	 * @return 是否空白符
 	 * @since 3.0.6
+	 * @deprecated 请使用{@link CharUtil#isBlankChar(int)}
 	 */
+	@Deprecated
 	public static boolean isBlankChar(int c) {
-		return Character.isWhitespace(c) || Character.isSpaceChar(c);
+		return Character.isWhitespace(c) || Character.isSpaceChar(c) || c == '\ufeff' || c == '\u202a';
 	}
 
 	/**