+++ hutool-core/src/test/java/cn/hutool/core/date/DateUtilTest.java
@@ -113,12 +113,17 @@ public class DateUtilTest {
 		Date newDate = DateUtil.offset(date, DateField.DAY_OF_MONTH, 2);
 		Assert.assertEquals("2017-03-03 22:33:23", newDate.toString());
 		
-		//常用偏移
+		//偏移天
 		DateTime newDate2 = DateUtil.offsetDay(date, 3);
 		Assert.assertEquals("2017-03-04 22:33:23", newDate2.toString());
-		//常用偏移
+		
+		//偏移小时
 		DateTime newDate3 = DateUtil.offsetHour(date, -3);
 		Assert.assertEquals("2017-03-01 19:33:23", newDate3.toString());
+		
+		//偏移月
+		DateTime offsetMonth = DateUtil.offsetMonth(date, -1);
+		Assert.assertEquals("2017-02-01 22:33:23", offsetMonth.toString());
 	}
 	
 	@Test