+++ hutool-core/src/test/java/cn/hutool/core/io/FileUtilTest.java
@@ -9,9 +9,7 @@ import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.file.LineSeparator;
-import cn.hutool.core.lang.Console;
 import cn.hutool.core.util.CharsetUtil;
 
 /**
@@ -20,40 +18,34 @@ import cn.hutool.core.util.CharsetUtil;
  * @author Looly
  */
 public class FileUtilTest {
-	
+
 	@Test
-	public void getAbsolutePathTest(){
+	public void getAbsolutePathTest() {
 		String absolutePath = FileUtil.getAbsolutePath("LICENSE-junit.txt");
 		String absolutePath2 = FileUtil.getAbsolutePath(absolutePath);
 		Assert.assertNotNull(absolutePath2);
 		Assert.assertEquals(absolutePath, absolutePath2);
 	}
-	
+
 	@Test
 	public void getAbsolutePathTest2() {
 		String path = FileUtil.getAbsolutePath("中文.xml");
 		Assert.assertTrue(path.contains("中文.xml"));
 	}
-	
-	@Test
-	public void getAbsolutePathTest3() {
-		String path = FileUtil.getAbsolutePath("../../LICENSE-junit.txt");
-		Console.log(path);
-	}
-	
+
 	@Test
 	@Ignore
 	public void touchTest() {
 		FileUtil.touch("d:\\tea\\a.jpg");
 	}
-	
+
 	@Test
 	public void delTest() {
-		//删除一个不存在的文件，应返回false
+		// 删除一个不存在的文件，应返回false
 		boolean result = FileUtil.del("e:/Hutool_test_3434543533409843.txt");
 		Assert.assertFalse(result);
 	}
-	
+
 	@Test
 	@Ignore
 	public void renameTest() {
@@ -66,34 +58,34 @@ public class FileUtilTest {
 		File destFile = FileUtil.file("hutool.copy.jpg");
 
 		FileUtil.copy(srcFile, destFile, true);
-		
+
 		Assert.assertTrue(destFile.exists());
 		Assert.assertEquals(srcFile.length(), destFile.length());
 	}
-	
+
 	@Test
 	public void equlasTest() {
-		//源文件和目标文件都不存在
+		// 源文件和目标文件都不存在
 		File srcFile = FileUtil.file("d:/hutool.jpg");
 		File destFile = FileUtil.file("d:/hutool.jpg");
-		
+
 		boolean equals = FileUtil.equals(srcFile, destFile);
 		Assert.assertTrue(equals);
-		
-		//源文件存在，目标文件不存在
+
+		// 源文件存在，目标文件不存在
 		File srcFile1 = FileUtil.file("hutool.jpg");
 		File destFile1 = FileUtil.file("d:/hutool.jpg");
-		
+
 		boolean notEquals = FileUtil.equals(srcFile1, destFile1);
 		Assert.assertFalse(notEquals);
 	}
-	
+
 	@Test
 	@Ignore
 	public void convertLineSeparatorTest() {
 		FileUtil.convertLineSeparator(FileUtil.file("d:/aaa.txt"), CharsetUtil.CHARSET_UTF_8, LineSeparator.WINDOWS);
 	}
-	
+
 	@Test
 	public void normalizeTest() {
 		Assert.assertEquals("/foo/", FileUtil.normalize("/foo//"));
@@ -114,12 +106,12 @@ public class FileUtilTest {
 		Assert.assertEquals("bar", FileUtil.normalize("../../bar"));
 		Assert.assertEquals("C:/bar", FileUtil.normalize("/C:/bar"));
 	}
-	
+
 	@Test
 	public void normalizeClassPathTest() {
 		Assert.assertEquals("", FileUtil.normalize("classpath:"));
 	}
-	
+
 	@Test
 	public void doubleNormalizeTest() {
 		String normalize = FileUtil.normalize("/aa/b:/c");
@@ -127,19 +119,19 @@ public class FileUtilTest {
 		Assert.assertEquals("/aa/b:/c", normalize);
 		Assert.assertEquals(normalize, normalize2);
 	}
-	
+
 	@Test
 	public void subPathTest() {
 		Path path = Paths.get("d:/aaa/bbb/ccc/ddd/eee/fff");
-		
+
 		Path subPath = FileUtil.subPath(path, 5, 4);
 		Assert.assertEquals("eee", subPath.toString());
 		subPath = FileUtil.subPath(path, 0, 1);
 		Assert.assertEquals("aaa", subPath.toString());
 		subPath = FileUtil.subPath(path, 1, 0);
 		Assert.assertEquals("aaa", subPath.toString());
-		
-		//负数
+
+		// 负数
 		subPath = FileUtil.subPath(path, -1, 0);
 		Assert.assertEquals("aaa/bbb/ccc/ddd/eee", subPath.toString().replace('\\', '/'));
 		subPath = FileUtil.subPath(path, -1, Integer.MAX_VALUE);
@@ -149,35 +141,35 @@ public class FileUtilTest {
 		subPath = FileUtil.subPath(path, -2, -3);
 		Assert.assertEquals("ddd", subPath.toString());
 	}
-	
+
 	@Test
 	public void subPathTest2() {
 		String subPath = FileUtil.subPath("d:/aaa/bbb/", "d:/aaa/bbb/ccc/");
 		Assert.assertEquals("ccc/", subPath);
-		
+
 		subPath = FileUtil.subPath("d:/aaa/bbb", "d:/aaa/bbb/ccc/");
 		Assert.assertEquals("ccc/", subPath);
-		
+
 		subPath = FileUtil.subPath("d:/aaa/bbb", "d:/aaa/bbb/ccc/test.txt");
 		Assert.assertEquals("ccc/test.txt", subPath);
-		
+
 		subPath = FileUtil.subPath("d:/aaa/bbb/", "d:/aaa/bbb/ccc");
 		Assert.assertEquals("ccc", subPath);
-		
+
 		subPath = FileUtil.subPath("d:/aaa/bbb", "d:/aaa/bbb/ccc");
 		Assert.assertEquals("ccc", subPath);
-		
+
 		subPath = FileUtil.subPath("d:/aaa/bbb", "d:/aaa/bbb");
 		Assert.assertEquals("", subPath);
-		
+
 		subPath = FileUtil.subPath("d:/aaa/bbb/", "d:/aaa/bbb");
 		Assert.assertEquals("", subPath);
 	}
-	
+
 	@Test
 	public void getPathEle() {
 		Path path = Paths.get("d:/aaa/bbb/ccc/ddd/eee/fff");
-		
+
 		Path ele = FileUtil.getPathEle(path, -1);
 		Assert.assertEquals("fff", ele.toString());
 		ele = FileUtil.getPathEle(path, 0);
@@ -187,15 +179,15 @@ public class FileUtilTest {
 		ele = FileUtil.getPathEle(path, -6);
 		Assert.assertEquals("aaa", ele.toString());
 	}
-	
+
 	@Test
 	public void listFileNamesTest() {
 		List<String> names = FileUtil.listFileNames("classpath:");
 		Assert.assertTrue(names.contains("hutool.jpg"));
-		
+
 		names = FileUtil.listFileNames("");
 		Assert.assertTrue(names.contains("hutool.jpg"));
-		
+
 		names = FileUtil.listFileNames(".");
 		Assert.assertTrue(names.contains("hutool.jpg"));
 	}