+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * 【poi】       ExcelWriter.merge方法加入重载，可选是否加入默认标题样式
 * 【poi】       ExcelSaxReader改进按照流读取工作簿的构造，使之对于mark不支持的流也可解析
 * 【cron】     添加updatePattern方法，可更新Task执行时间规则(感谢@【上海】嘿)
+* 【cache】   添加get方法支持可选的是否更新lastAccess时间（issue#IISC4@Gitee）
+* 【core】     StrUtil增加isNullOrUndefined、isEmptyOrUndefined、isBlankOrUndefined方法（issue#IIR44@Gitee）
+* 【core】     isBlankChar方法迁移到CharUtil中
 
 ### Bug修复
 * 【poi】       修复ExcelWriter合并单元格后样式失效问题